+++ changelog.md
@@ -4,3 +4,4 @@
 ## v0.0.1 - 2017-01-20
 
 - Fonctionnalités des scripts/skins  
+- Home page : tri/filtre/recherche