+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## v0.0.1 - 2017-01-20
+## v0.0.1 - 2017-02
 
 - Fonctionnalités des scripts/skins  
 - Home page : tri/filtre/recherche